refactor(workspace): extract connectDoc, the shared doc wiring
A workspace is a tree of Y.Docs that all speak to one relay under one owner: the
root (tables/kv + actions) and every body child doc (an attach* layout). Both
want identical wiring (attachLocalStorage + openCollaboration), which every
app's browser.ts had hand-copied as a local `wire` helper and createChildDocs
inlined a third time.

Extract `connectDoc(ydoc, config, { actions })` as that one primitive and route
createChildDocs through it. `ChildDocConnection` becomes an alias of the shared
`ConnectionConfig`, retiring when `defineWorkspace(...).open(connection)` lands.
No behavior change.
